requirements.txt
torch==2.9.0 
torchvision==0.24.0 
Flask 
Pillow 
gunicorn 
